Lottery post fixes (#22)
* Master (#21)

* Method to beat the Bingo scratch lottery game

Using image processing and optical character recognition winning Bingo lottery tickets may be recognized prior to scratching off the hidden numbers. The method improves the expected value of the game sufficiently to make it worth playing.
By counting the number of times each number 0-99 shows up on any of the Bingo cards we can tell with 90% accuracy if there is a winning solution.

* Method to beat the Bingo scratch lottery game

Using image processing and optical character recognition winning Bingo lottery tickets may be recognized prior to scratching off the hidden numbers. The method improves the expected value of the game sufficiently to make it worth playing.

* Fix fraction lines color in Katex

* Bump TailwindCSS dependency version

* Basic blogpost <table> styles

* Make /docs available

* Remove duplicate lottery post
- made sure it didn't contain unique data by doing a file comparison

* Lottery post fixes and improvements
- fixes images and cleans up filenames
- right align table columns and styling
- fix OCR cards layout

Co-authored-by: XerxesZorgon <the.one.true.zorgon.x@gmail.com>