Using the Geogebra CAS for integrals
Integration is the reverse process of the derivative, and can be interpreted as finding the area under a curve between two fixed points along the x-axis. We show how to use the Geogebra CAS to sovle for the integral, and then use the CAS to find a solution to a problem proposed by the Classical Mathematics Group.